Delete svr_jaya_desharnais.ipynb
This is the third week this file has been online on GitHub, I haven't submitted it to my supervisor yet and I fear someone might copy it and publish it before I do... Making it hard or impossible for me to claim it or submit it to my department as my final year school project. I'm leaving this commit message as some sort of receipt to show that I uploaded the files 3 weeks back and deleted them today, just in case I'll need to prove that this is my work.
Maybe I'm just overreacting, who knows.